+++ library/book/form.py
@@ -1,6 +1,7 @@
 from django import forms
-from .models import author, genre
+from .models import author, genre, publisher, book
 
+#region ================================= AUTHOR AREA ==========================================================================
 
 class form_create_author(forms.ModelForm):
     first_name = forms.CharField(label= "First Name")
@@ -26,6 +27,10 @@ class form_edit_author(forms.ModelForm):
     class Meta:
         model = author
         fields = ['first_name', 'middle_name', 'last_name', 'is_active']
+#endregion ================================= AUTHOR AREA ==========================================================================
+
+
+#region ================================= GENRE AREA ==========================================================================
 
 class form_create_genre(forms.ModelForm):
 
@@ -45,3 +50,90 @@ class form_edit_genre(forms.ModelForm):
     class Meta:
         model = genre
         fields = ['name', 'is_active']
+#endregion ================================= GENRE AREA ==========================================================================
+
+
+#region ================================= PUBLISHER AREA ==========================================================================
+
+class form_create_publisher(forms.ModelForm):
+
+    class Meta:
+        model = publisher
+        fields = ['name']
+
+class form_edit_publisher(forms.ModelForm):
+    ACTIVE_CHOICES = [
+        (False, "Not Active"),
+        (True, 'Active'),
+    ]
+
+    is_active = forms.ChoiceField(choices=ACTIVE_CHOICES, label='Select Activate', required=False,
+                                   widget=forms.Select(attrs={'class': 'form-control'}))
+
+    class Meta:
+        model = publisher
+        fields = ['name', 'is_active']
+#endregion ================================= PUBLISHER AREA ==========================================================================
+
+
+#region ================================= BOOK AREA ==========================================================================
+
+class form_create_book(forms.ModelForm):
+    tittle = forms.CharField(label= "Tittle")
+    author = forms.ModelChoiceField(queryset=author.objects.all(), label="Author")
+    genre = forms.ModelChoiceField(queryset=genre.objects.all(), label= "Genre")
+    publisher = forms.ModelChoiceField(queryset=publisher.objects.all(), label= "Publisher")
+    
+    total_page = forms.IntegerField(required=False, label= "Total Page")
+    isbn = forms.CharField(required=False, label= "ISBN")
+    published_date = forms.DateField(required=False, label= "Published Date", widget=forms.DateInput(attrs={'type': 'date'}))
+
+    # python constructor
+    def __init__(self, *args, **kwargs):
+        super(form_create_book, self).__init__(*args, **kwargs)
+        # Customize the label of each choice in the dropdown
+        self.fields['author'].label_from_instance = self.label_from_author_instance
+
+    def label_from_author_instance(self, obj):
+        # Customize how each author instance is displayed in the dropdown
+        return f"{obj.first_name} {obj.middle_name} {obj.last_name}".strip()
+
+    class Meta:
+        model = book
+        fields = ['tittle', 'author', 'genre', 'publisher', 'total_page', 'isbn', 'published_date']
+
+
+
+
+class form_edit_book(forms.ModelForm):
+    ACTIVE_CHOICES = [
+        (False, "Not Active"),
+        (True, 'Active'),
+    ]
+
+    is_active = forms.ChoiceField(choices=ACTIVE_CHOICES, label='Select Activate', required=False,
+                                   widget=forms.Select(attrs={'class': 'form-control'}))
+    
+    tittle = forms.CharField(label= "Tittle")
+    author = forms.ModelChoiceField(queryset=author.objects.all(), label="Author")
+    genre = forms.ModelChoiceField(queryset=genre.objects.all(), label= "Genre")
+    publisher = forms.ModelChoiceField(queryset=publisher.objects.all(), label= "Publisher")
+    
+    total_page = forms.IntegerField(required=False, label= "Total Page")
+    isbn = forms.CharField(required=False, label= "ISBN")
+    published_date = forms.DateField(required=False, label= "Published Date", widget=forms.DateInput(attrs={'type': 'date'}))
+    
+    # python constructor
+    def __init__(self, *args, **kwargs):
+        super(form_edit_book, self).__init__(*args, **kwargs)
+        # Customize the label of each choice in the dropdown
+        self.fields['author'].label_from_instance = self.label_from_author_instance
+
+    def label_from_author_instance(self, obj):
+        # Customize how each author instance is displayed in the dropdown
+        return f"{obj.first_name} {obj.middle_name} {obj.last_name}".strip()
+    
+    class Meta:
+        model = book
+        fields = ['tittle', 'author', 'genre', 'publisher', 'total_page', 'isbn', 'published_date', 'is_active']
+#endregion ================================= BOOK AREA ==========================================================================