+++ library/book/models.py
@@ -31,7 +31,7 @@ class book(models.Model):
     publisher = models.ForeignKey(publisher, on_delete=models.CASCADE)
     total_page = models.IntegerField(null= True)
     isbn = models.CharField(max_length=50, null= True)
-    published_date = models.DateField(max_length=50, null= True)
+    published_date = models.DateField(null= True)
     is_active = models.BooleanField(default= True, help_text= "0 not active; 1 active")
 
     def __str__(self):