+++ library/book/urls.py
@@ -11,6 +11,16 @@ urlpatterns = [
     path("genre/create/", views.genre_create, name = "genre_create"),
     path("genre/<int:id>/", views.genre_update, name = "genre_update"),
     path("genre/<int:id>/delete/", views.genre_delete, name = "genre_delete"),
+
+    path("publisher/", views.publisher_read, name = "publisher_read"),
+    path("publisher/create/", views.publisher_create, name = "publisher_create"),
+    path("publisher/<int:id>/", views.publisher_update, name = "publisher_update"),
+    path("publisher/<int:id>/delete/", views.publisher_delete, name = "publisher_delete"),
+
+    path("book/", views.book_read, name = "book_read"),
+    path("book/create/", views.book_create, name = "book_create"),
+    path("book/<int:id>/", views.book_update, name = "book_update"),
+    path("book/<int:id>/delete/", views.book_delete, name = "book_delete"),
 ]
 
 '''