+++ library/book/views.py
@@ -1,7 +1,8 @@
 from django.shortcuts import render, get_object_or_404, redirect, HttpResponseRedirect
 from django.contrib.auth.decorators import login_required
-from .models import author as auth, genre
+from .models import author as auth, genre, publisher, book
 from .form import form_create_author, form_edit_author, form_create_genre, form_edit_genre
+from .form import form_create_publisher, form_edit_publisher, form_create_book, form_edit_book
 # Create your views here.
 
 #region ================================= AUTHOR AREA ==========================================================================
@@ -120,3 +121,117 @@ def genre_delete(response, id):
         data.delete()
         return redirect('genre_read') 
 #endregion ================================= GENRE AREA ==========================================================================
+
+#region ================================= PUBLISHER AREA ==========================================================================
+
+@login_required()
+def publisher_read(response):
+    data = publisher.objects.all()
+    return render(response, "book/publisher_read.html", {"data":data})
+
+@login_required()
+def publisher_create(response):
+    #Authorisation
+    author = ['superadmin', 'admin']
+    if response.user.role not in author:
+        return HttpResponseRedirect('/')
+    
+    if response.method == "POST":
+        form = form_create_publisher(response.POST)
+        if form.is_valid():
+            instance = form.save(commit=False)  # Prevent premature saving
+            instance.save()
+            return redirect('publisher_read')
+    else:
+        form = form_create_publisher()
+    return render(response, "book/publisher_create.html", {"form":form})
+
+@login_required()
+def publisher_update(response, id):
+    data = get_object_or_404(publisher, id=id)
+
+    #Authorisation (Note: Change more specific)
+    author = ['superadmin', 'admin']
+    if response.user.role not in author:
+        return HttpResponseRedirect('/')
+    
+    form = form_edit_publisher(response.POST or None, instance=data)
+    if response.method =="POST":
+        if form.is_valid():
+            instance = form.save(commit=False)  # Prevent premature saving
+
+            instance.save()
+            return redirect('publisher_read')
+    
+    return render(response, "book/publisher_edit.html", {"form":form, "data":data})
+
+@login_required()
+def publisher_delete(response, id):
+    data = get_object_or_404(publisher, id=id)
+
+    #Authorisation
+    author = ['superadmin']
+    if response.user.role not in author:
+        return HttpResponseRedirect('/')
+    else:
+        data.delete()
+        return redirect('publisher_read') 
+
+#endregion ================================= PUBLISHER AREA ==========================================================================
+
+#region ================================= BOOK AREA ==========================================================================
+
+@login_required()
+def book_read(response):
+    data = book.objects.all()
+    return render(response, "book/book_read.html", {"data":data})
+
+@login_required()
+def book_create(response):
+    #Authorisation
+    author = ['superadmin', 'admin']
+    if response.user.role not in author:
+        return HttpResponseRedirect('/')
+    
+    if response.method == "POST":
+        form = form_create_book(response.POST)
+        if form.is_valid():
+            instance = form.save(commit=False)  # Prevent premature saving
+            instance.save()
+            return redirect('book_read')
+    else:
+        form = form_create_book()
+    return render(response, "book/book_create.html", {"form":form})
+
+@login_required()
+def book_update(response, id):
+    data = get_object_or_404(book, id=id)
+
+    #Authorisation (Note: Change more specific)
+    author = ['superadmin', 'admin']
+    if response.user.role not in author:
+        return HttpResponseRedirect('/')
+    
+    form = form_edit_book(response.POST or None, instance=data)
+    if response.method =="POST":
+        if form.is_valid():
+            instance = form.save(commit=False)  # Prevent premature saving
+
+            instance.save()
+            return redirect('book_read')
+    
+    return render(response, "book/book_edit.html", {"form":form, "data":data})
+
+@login_required()
+def book_delete(response, id):
+    data = get_object_or_404(book, id=id)
+
+    #Authorisation
+    author = ['superadmin']
+    if response.user.role not in author:
+        return HttpResponseRedirect('/')
+    else:
+        data.delete()
+        return redirect('book_read')
+#endregion ================================= BOOK AREA ==========================================================================
+